Memindahkan function verification ke LogisticVerificationController
